+++ src/helpers/CategoriesFilterHelper.ts
